#66 編集した画像が表示できないバグを修正し、軽微なリファクタリングも行った

+++ app/Http/Controllers/Admin/ProfileController.php
@@ -35,12 +35,11 @@ class ProfileController extends Controller
         // (可読性を考え画像を削除する場合の処理を先にした)
         if (strcmp($request->remove, 'true') == 0) {
             Storage::disk('s3')->delete('user_images/' . $user_data->image_path);
-            // Storage::delete('public/user_image/' . $user_data->image_path); // ローカル環境での記述
             $user_data->user_image_path = null;
-        } elseif (isset($users_form['image'])) {
+        } elseif (isset($users_form['image_file'])) {
 
             // 新しい画像ファイルとファイル名を取得
-            $posted_image = $request->file('image');
+            $posted_image = $request->file('image_file');
 
             // 画像をリサイズしてjpgにencodeする(InterventionImageのImageファサードを使用)
             $resized_image = Image::make($posted_image)->fit(200, 200)->encode('jpg'); // CSSで円形にする
@@ -52,10 +51,11 @@ class ProfileController extends Controller
             $image_hash = md5($resized_image->__toString());
             $image_name = "{$image_hash}.jpg";
 
-            // 現在設定中の画像を削除し、加工した新しい画像を保存する
-            Storage::disk('s3')->delete('user_images/' . $user_data->image_path);
+            // 現在設定中の画像があれば削除し、加工した新しい画像を保存する
+            if (isset($user_data->image_path)) {
+                Storage::disk('s3')->delete('user_images/' . $user_data->image_path);
+            }
             Storage::disk('s3')->put('user_images/' . $image_name, $resized_image, 'public');
-            // Storage::put('public/user_image/' . $image_name, $resized_image); // ローカル環境での記述
 
             // S3のファイルのURLを取得してuser_image_pathカラムに書き込む
             $user_data->user_image_path = Storage::disk('s3')->url('user_images/' . $image_name);
@@ -70,11 +70,14 @@ class ProfileController extends Controller
     // deleteアクション
     public function delete(Request $request)
     {
-        // Modelからデータを取得して削除する
+        // Modelからデータを取得する(idで検索)
         $user_data = User::find($request->id);
-        Storage::disk('s3')->delete('user_images/' . $user_data->image_path); // S3のファイルを削除
-        // Storage::delete('public/image/' . $user_data->image_path); // ローカル環境での記述
-        $user_data->delete(); // テーブルのファイルパスを削除
+
+        // 画像パスがあればS3の画像ファイルを削除し、テーブルのデータを削除する
+        if (isset($user_data->image_path)) {
+            Storage::disk('s3')->delete('user_images/' . $user_data->image_path);
+        }
+        $user_data->delete();
 
         return redirect('admin/users');
     }
